requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic[email]==2.7.4
pydantic-settings==2.4.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
requests==2.32.3
openpyxl==3.1.5
reportlab==4.2.5
httpx==0.27.0
PyYAML==6.0.1
pytest==8.2.2
anthropic==0.49.0
